+++ src/modules/create_graph/data_parser/data_handler.py
@@ -1,24 +1,108 @@
+import sys
+
 import xml.sax
 import networkx as nx
+from src.modules.create_graph.utils import utils
 from src.modules.create_graph.data_parser.parse_osm import OsmParsers
+from src.modules.create_graph.data_parser.parse_posts import PostHandler
+from src.modules.create_graph.pojo.search_node import SearchNode
 
 class DataHandler():
 
-    def __init__(self):
-        # create_graph an XMLReader
+
+
+    def align_nodes_and_posts(self, nodesL):
+        '''
+        Align post offices with nodes
+
+        :param posts:
+        :param nodesL:
+        :return:
+        '''
+
+        postsNodes = []
+        for post in self.posts:
+            minDist = sys.maxsize
+            nodeKey = ""
+            for key, node in nodesL.items():
+                tmpDist = utils.calcDistance(post.latitude, post.longitude, node.lat, node.lon)
+                if minDist > tmpDist:
+                    minDist = tmpDist
+                    nodeKey = key
+
+            tmpNode = nodesL[nodeKey]
+            if utils.calcDistance(post.latitude, post.longitude, tmpNode.lat, tmpNode.lon) < 1:
+                print(nodeKey)
+                print(post.address)
+                print(utils.calcDistance(post.latitude, post.longitude, tmpNode.lat, tmpNode.lon))
+                postsNodes.append(nodeKey)
+                tmpNode.add_post(post.address)
+                nodesL[nodeKey] = tmpNode
+        return (nodesL, postsNodes)
+
+
+
+
+    def retrive_posts(self, post_path):
+        post_handler = PostHandler()
+        self.posts = post_handler.read_postal_offices(post_path)
+
+
+
+    def retrieve_road_topology(self, osm_path, post_path):
+
+        self.retrive_posts(post_path)
 
         parser = xml.sax.make_parser()
-        # turn off namepsaces
         parser.setFeature(xml.sax.handler.feature_namespaces, 0)
 
-        # override the default ContextHandler
+        # parse osm files
         handler = OsmParsers()
         parser.setContentHandler(handler)
-
-        parser.parse("data/test_export.osm")
+        parser.parse(osm_path)
 
         self.ways = handler.ways
         self.nodes = handler.nodes
+
+        nodes_filtered = {}
+        #in this step we remove nodes which are not connected to road
+        #with this step we want to avoid that algoritm will hang on.
+        for way in self.ways:
+            for id in way.ids:
+                nodes_filtered[id] = self.nodes[id]
+
+        nodesDict = {}
+        edgesDict = {}
+
+        #alignement
+        (roadNodesAnotated, postsNodes) =  self.align_nodes_and_posts(nodes_filtered)
+
+        i = 1
+        #generate ides for post offices (generic)
+        for key, node in roadNodesAnotated.items():
+            if node.post:
+                nodesDict[node.id] = SearchNode(node.id, i, node.post)
+                i = i + 1
+            else:
+                nodesDict[node.id] = SearchNode(node.id)
+
+
+        for key, node in roadNodesAnotated.items():
+            edgesDict[node.id] = {}
+
+        for way in self.ways:
+            tmpD = edgesDict[way.ids[0]]
+            tmpD[way.ids[1]] = {"weight": way.distance}
+            edgesDict[way.ids[0]] = tmpD
+
+        self.modified_ways = edgesDict
+        self.modified_nodes = nodesDict
+
+
+    def __init__(self, osm_path, posts_path):
+        # create_graph an XMLReader
+        self.retrieve_road_topology(osm_path, posts_path)
+
 
     def graph_viz(self):
         G = nx.Graph()
