+++ src/modules/create_graph/data_parser/parse_posts.py
@@ -1,7 +1,6 @@
-import sys
+
 import csv
 from src.modules.create_graph.pojo.post import Post
-from src.modules.create_graph.utils import utils
 
 class PostHandler:
 
@@ -14,11 +13,11 @@ class PostHandler:
         except ValueError:
             return None
 
-    def read_postal_offices(self):
+    def read_postal_offices(self, post_path):
         ''' Postal offices are read from csv file and than added to array
         '''
         self.posts = []
-        with open('config/List of Postal Offices (geographical location).csv') as csv_file:
+        with open(post_path) as csv_file:
             csv_reader = csv.reader(csv_file, delimiter=',')
 
             for row in csv_reader:
@@ -28,32 +27,3 @@ class PostHandler:
                     self.posts.append(post)
 
         return self.posts
-
-    def align_nodes_and_posts(self, nodesL):
-        '''
-        Align post offices with nodes
-
-        :param posts:
-        :param nodesL:
-        :return:
-        '''
-        posts = self.read_postal_offices()
-        postsNodes = []
-        for post in posts:
-            minDist = sys.maxsize
-            nodeKey = ""
-            for key, node in nodesL.items():
-                tmpDist = utils.calcDistance(post.latitude, post.longitude, node.lat, node.lon)
-                if minDist > tmpDist:
-                    minDist = tmpDist
-                    nodeKey = key
-
-            tmpNode = nodesL[nodeKey]
-            if utils.calcDistance(post.latitude, post.longitude, tmpNode.lat, tmpNode.lon) < 1:
-                print(nodeKey)
-                print(post.address)
-                print(utils.calcDistance(post.latitude, post.longitude, tmpNode.lat, tmpNode.lon))
-                postsNodes.append(nodeKey)
-                tmpNode.add_post(post.address)
-                nodesL[nodeKey] = tmpNode
-        return (nodesL, postsNodes)
